% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha_hcl28}
\alias{alpha_hcl28}
\title{Calculates Cronbach's alpha based on HCL-28 model}
\usage{
alpha_hcl28(data)
}
\arguments{
\item{data}{Dataset containing HCL-32 items.}
}
\value{
Cronbach's alpha coefficient calculated based on Pearson's correlation matrix in the full dataset, active factor and risk-taking factor.
}
\description{
Calculates Cronbach's alpha based on HCL-28 model
}
