% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{create_models}
\alias{create_models}
\title{Create three models that can be used to fit lavaan models for HCL-32}
\usage{
create_models()
}
\value{
A list with three models on lavaan package syntax.
}
\description{
Create three models that can be used to fit lavaan models for HCL-32
}
