% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{fit_models}
\alias{fit_models}
\title{Fit CFA models on given data using WLSMV estimator}
\usage{
fit_models(data, models)
}
\arguments{
\item{data}{Data frame or tibble containing the HCL-32 items.}

\item{models}{Models to be fitted on data.}
}
\value{
A list containing the fitted models.
}
\description{
Fit CFA models on given data using WLSMV estimator
}
