% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{ord_alpha}
\alias{ord_alpha}
\title{Calculates ordinal alpha}
\usage{
ord_alpha(data)
}
\arguments{
\item{data}{Dataset containing HCL-32 items.}
}
\value{
Ordinal alpha coefficient calculated based on tetrachoric correlation matrix.
}
\description{
Calculates ordinal alpha
}
