% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{summarize_fit}
\alias{summarize_fit}
\title{Summarize fitted models performance metrics}
\usage{
summarize_fit(fits)
}
\arguments{
\item{fits}{A list containing fitted models.}
}
\value{
Summarized metrics for each model as a list. Based on lavaan output.
}
\description{
Summarize fitted models performance metrics
}
