% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{summarize_parameters}
\alias{summarize_parameters}
\title{Summarize fitted models metrics}
\usage{
summarize_parameters(fits)
}
\arguments{
\item{fits}{A list containing fitted models.}
}
\value{
Summarized parameters for each model as a list. Based on lavaan output.
}
\description{
Summarize fitted models metrics
}
